Refactor login and signup to same shared component for styling
